+++ FlexColorPicker/Classes/ColorPickerControllerProtocol.swift
@@ -28,7 +28,7 @@
 
 import UIKit
 
-/// Protocol implemented by ViewControllers that contain controls to pick color.
+/// Protocol implemented by controllers that connect and synchronize *color controls* (`UIView`s, see protocol `ColorControl`) used to pick color. These can be view controllers that also display those *color controls*.
 public protocol ColorPickerControllerProtocol: class {
     /// Delegate notified when selected color is changed or when user confirms currently selected color.
     var delegate: ColorPickerDelegate? { get set }