+++ FlexColorPicker/Classes/CustomColorPickerViewController.swift
@@ -160,7 +160,9 @@ open class CustomColorPickerViewController: UIViewController, ColorPickerControl
     }
 
     open override func viewDidLoad() {
-        if view.backgroundColor == nil {
+        if #available(iOS 13.0, *) {
+            view.backgroundColor = .systemBackground
+        } else {
             view.backgroundColor = .white
         }
         super.viewDidLoad()