+++ FlexColorPicker/Classes/Views/ColorPickerThumbView.swift
@@ -89,14 +89,14 @@ open class ColorPickerThumbView: UIViewWithCommonInit {
         addAutolayoutFillingSubview(borderView)
         addAutolayoutFillingSubview(colorView, edgeInsets: UIEdgeInsets(top: wideBorderWidth, left: wideBorderWidth, bottom: wideBorderWidth, right: wideBorderWidth))
         addAutolayoutCentredView(percentageLabel)
-        borderView.borderColor = UIColor(named: "BorderColor")
+        borderView.viewBorderColor = UIColor.colorPickerBorderColor
         borderView.borderWidth = 1 / UIScreen.main.scale
         percentageLabel.font = percentageTextFont
-        percentageLabel.textColor = UIColor(named: "LabelTextsColor")
+        percentageLabel.textColor = UIColor.colorPickerLabelTextColor
         percentageLabel.textAlignment = .center
         percentageLabel.alpha = 0
         clipsToBounds = false // required for the text label to be displayed ourside of bounds
-        borderView.backgroundColor = UIColor(named: "ThumbViewWideBorderColor")
+        borderView.backgroundColor = UIColor.colorPickerThumbViewWideBorderColor
         (borderView as? LimitedGestureCircleView)?.delegate = delegate
         (colorView as? LimitedGestureCircleView)?.delegate = delegate
         setColor(color, animateBorderColor: false)
@@ -148,8 +148,8 @@ extension ColorPickerThumbView {
     }
 
     private func setWideBorderColors(_ isDark: Bool) {
-        self.borderView.borderColor = UIColor(named: isDark ? "BorderColor" : "LightBorderColor")
-        self.borderView.backgroundColor = UIColor(named: isDark ? "ThumbViewWideBorderDarkColor" : "ThumbViewWideBorderColor")
+        self.borderView.viewBorderColor = isDark ? UIColor.colorPickerBorderColor : UIColor.colorPickerLightBorderColor
+        self.borderView.backgroundColor = isDark ? UIColor.colorPickerThumbViewWideBorderDarkColor : UIColor.colorPickerThumbViewWideBorderColor
     }
 }